requirements.txt
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-experimental
pdfplumber
sentence-transformers
qdrant-client
streamlit
python-dotenv